Refactor `BidderProfile`
